sentence_transformer/requirements.txt
elasticsearch
sentence-transformers
bert-extractive-summarizer
neuralcoref
spacy
pandas
rake-nltk
trafilatura
argparse
tqdm
nltk